feat: exclude certain operations
Input data is processed for `don't()` flags that disable processing of any subsequent `mul()` operations until encountering a `do()` flag that once again toggles their processing back on.

This commit also includes a refactoring of code to further simplify the amount of function calls and logic of the program, at the possible expense of using a bit more memory by loading in the entire input data instead of a generator to load the data line-by-line.